Updated HTML/CSS for Athenium Challenge.

+++ CodingExercise/calculations.js
@@ -15,7 +15,7 @@ function calculateHours() {
 	if(employeeName == "" || employeeName == null || hourlyWage == "" || hourlyWage == null || weekDates == "" || weekDates == null || mondayHours == "" || mondayHours == null || tuesdayHours == "" || tuesdayHours == null || wednesdayHours == "" || wednesdayHours == null || thursdayHours === "" || thursdayHours == null || fridayhours === "" || fridayHours == null) {
 		window.alert("Make sure that all of the fields are filled in before you click 'Calculate'!");
 	} else {
-		var hourlyWageNumber = parseInt(hourlyWage);
+		var hourlyWageNumber = parseFloat(hourlyWage);
 
 		var text = "Hourly wage is: " + hourlyWageNumber;
